backend/requirements.txt
# Django Backend for MindSpark AI Quiz

## Requirements
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-generativeai==0.3.2
django-jazzmin
